% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_installed}
\alias{check_installed}
\title{Check that a package installed}
\usage{
check_installed(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\description{
Checks that a package is installed, throws
an error if it is not.
}
\keyword{internal}
