% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_urls}
\alias{check_urls}
\title{Check URLs}
\usage{
check_urls(fireblaze_ui)
}
\arguments{
\item{fireblaze_ui}{An object of class \code{\link{FireblazeUI}}.}
}
\description{
Check that tos and privacy policy urls are set.
}
