% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Config}
\usage{
create_config(api_key, project_id, auth_domain = NULL, overwrite = FALSE)
}
\arguments{
\item{api_key}{API key of your project.}

\item{project_id}{Id of your web project.}

\item{auth_domain}{Authentication domain, if omitted uses,
attempts to build firebase's default domain.}

\item{overwrite}{Whether to overwrite any existing configuration file.}
}
\value{
Path to file.
}
\description{
Create the configuration file necessary to running fireblaze.
}
\details{
Do not share this file with anyone.
}
