% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{use_fireblaze}
\alias{fireblaze_signin_ui}
\title{Dependencies}
\usage{
use_fireblaze(analytics = TRUE, firestore = FALSE)

fireblaze_signin_ui()
}
\arguments{
\item{analytics}{Whether to include analytics.}

\item{firestore}{Whether to include firestore.}
}
\description{
Include dependencies in your Shiny application.
\code{use_fireblaze} \emph{must} be included in
every application.
}
\section{Functions}{

\itemize{
\item{\code{use_fireblaze} Is required for every app that uses this package}
\item{\code{use_fireblaze_ui} Is required for applications that use \code{\link{FireblazeUI}}}
}
}

