% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{encryption}
\alias{encryption}
\alias{.enc}
\alias{.dec}
\title{Encryption}
\usage{
.enc(x)

.dec(x)
}
\arguments{
\item{x}{Value to encrypt.}
}
\description{
Encrypt and decrypt strings.
}
\keyword{internal}
