% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ai.R
\name{FirebaseAI}
\alias{FirebaseAI}
\title{Firebase AI}
\description{
Firebase AI

Firebase AI
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session}}{Shiny session}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FirebaseAI-new}{\code{FirebaseAI$new()}}
\item \href{#method-FirebaseAI-chat}{\code{FirebaseAI$chat()}}
\item \href{#method-FirebaseAI-clone}{\code{FirebaseAI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseAI-new"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseAI-new}{}}}
\subsection{Method \code{new()}}{
Initialize Firebase AI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseAI$new(session = shiny::getDefaultReactiveDomain())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{Shiny session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseAI-chat"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseAI-chat}{}}}
\subsection{Method \code{chat()}}{
Send a prompt to the AI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseAI$chat(id, prompt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the input to store the response in.}

\item{\code{prompt}}{The prompt to send to the AI.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseAI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseAI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseAI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
