% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Fireblaze}
\alias{Fireblaze}
\title{Fireblaze}
\description{
Use fireblaze to manage authentications.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session}}{A valid Shiny session.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fireblaze$new()}}
\item \href{#method-set_providers}{\code{Fireblaze$set_providers()}}
\item \href{#method-run}{\code{Fireblaze$run()}}
\item \href{#method-signed_in}{\code{Fireblaze$signed_in()}}
\item \href{#method-signing_in}{\code{Fireblaze$signing_in()}}
\item \href{#method-clone}{\code{Fireblaze$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$new(
  persistence = c("session", "none", "local"),
  config_path = "fireblaze.rds",
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{persistence}}{How the auth should persit: \code{none}, the user has to sign in at every visit,
\code{session} will only persist in current tab, \code{local} persist even when window is closed.}

\item{\code{config_path}}{Path to the configuration file as created by \code{\link{create_config}}.}

\item{\code{session}}{A valid shiny session.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise Fireblaze
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_providers"></a>}}
\if{latex}{\out{\hypertarget{method-set_providers}{}}}
\subsection{Method \code{set_providers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$set_providers(
  google = FALSE,
  facebook = FALSE,
  twitter = FALSE,
  github = FALSE,
  email = FALSE,
  phone = FALSE,
  anonymous = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{google, facebook, twitter, github, email, phone, anonymous}}{Set to \code{TRUE} the providers you want to use, at least one.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Define signin and login providers.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$run(flow = c("popup", "redirect"), helper = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{The signin flow to use, popup or redirect.}

\item{\code{helper}}{Wether to use accountchooser.com upon signing in or signing up with email,
the user will be redirected to the accountchooser.com website and will be able to select
one of their saved accounts. You can disable it by specifying the value below.}

\item{\code{...}}{Any other option to pass to Firebase UI.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Setup the signin form.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-signed_in"></a>}}
\if{latex}{\out{\hypertarget{method-signed_in}{}}}
\subsection{Method \code{signed_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$signed_in()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get Signed in User Info
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-signing_in"></a>}}
\if{latex}{\out{\hypertarget{method-signing_in}{}}}
\subsection{Method \code{signing_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$signing_in()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get Whether User is Currently Signing in
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
