% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{fireblaze_signin_ui}
\alias{fireblaze_signin_ui}
\title{Interface}
\usage{
fireblaze_signin_ui(hidden = FALSE)
}
\arguments{
\item{hidden}{Whether the UI is initialised shown or hidden.}
}
\description{
Interface
}
