% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-email-link.R
\name{FireblazeEmailLink}
\alias{FireblazeEmailLink}
\title{Email Link}
\description{
Email Link

Email Link
}
\section{Super class}{
\code{\link[fireblaze:Fireblaze]{fireblaze::Fireblaze}} -> \code{FireblazeEmailLink}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{email_verification}}{Email verification results}

\item{\code{email_sent}}{Email send results}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-config}{\code{FireblazeEmailLink$config()}}
\item \href{#method-send}{\code{FireblazeEmailLink$send()}}
\item \href{#method-verify}{\code{FireblazeEmailLink$verify()}}
\item \href{#method-get_email_sent}{\code{FireblazeEmailLink$get_email_sent()}}
\item \href{#method-get_email_verification}{\code{FireblazeEmailLink$get_email_verification()}}
\item \href{#method-clone}{\code{FireblazeEmailLink$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_in}{\code{fireblaze::Fireblaze$get_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_up">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_up}{\code{fireblaze::Fireblaze$get_signed_up()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="initialize">}\href{../../fireblaze/html/Fireblaze.html#method-initialize}{\code{fireblaze::Fireblaze$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="is_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-is_signed_in}{\code{fireblaze::Fireblaze$is_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="req_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-req_signed_in}{\code{fireblaze::Fireblaze$req_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="set_language_code">}\href{../../fireblaze/html/Fireblaze.html#method-set_language_code}{\code{fireblaze::Fireblaze$set_language_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="sign_out">}\href{../../fireblaze/html/Fireblaze.html#method-sign_out}{\code{fireblaze::Fireblaze$sign_out()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config"></a>}}
\if{latex}{\out{\hypertarget{method-config}{}}}
\subsection{Method \code{config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$config(url, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The link is handled in the web action widgets,
this is the deep link in the continueUrl query parameter.
Likely, your shiny application link.}

\item{\code{...}}{Any other parameter from the
\href{https://firebase.google.com/docs/auth/web/passing-state-in-email-actions?authuser=0#passing_statecontinue_url_in_email_actions}{official documentation}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Configure
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$send(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{Email to send verification to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Send email verification link.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify"></a>}}
\if{latex}{\out{\hypertarget{method-verify}{}}}
\subsection{Method \code{verify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$verify()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Verifies user came from email link.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_email_sent"></a>}}
\if{latex}{\out{\hypertarget{method-get_email_sent}{}}}
\subsection{Method \code{get_email_sent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$get_email_sent()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get whether email verification was correctly sent.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_email_verification"></a>}}
\if{latex}{\out{\hypertarget{method-get_email_verification}{}}}
\subsection{Method \code{get_email_verification()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$get_email_verification()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get whether email verification was correctly sent.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailLink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
