% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-email-password.R
\name{FireblazeEmailPassword}
\alias{FireblazeEmailPassword}
\title{Fireblaze}
\description{
Use fireblaze to manage authentications.
}
\section{Super class}{
\code{\link[fireblaze:Fireblaze]{fireblaze::Fireblaze}} -> \code{FireblazeEmailPassword}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-create}{\code{FireblazeEmailPassword$create()}}
\item \href{#method-sign_in}{\code{FireblazeEmailPassword$sign_in()}}
\item \href{#method-clone}{\code{FireblazeEmailPassword$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="initialize">}\href{../../fireblaze/html/Fireblaze.html#method-initialize}{\code{fireblaze::Fireblaze$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="sign_out">}\href{../../fireblaze/html/Fireblaze.html#method-sign_out}{\code{fireblaze::Fireblaze$sign_out()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-signed_in}{\code{fireblaze::Fireblaze$signed_in()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$create(email, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email, password}}{Credentials as entered by the user.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sign in with create
}

\subsection{Returns}{
\code{NULL} if successful, the error otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign_in"></a>}}
\if{latex}{\out{\hypertarget{method-sign_in}{}}}
\subsection{Method \code{sign_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$sign_in(email, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email, password}}{Credentials as entered by the user.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sign in with email
}

\subsection{Returns}{
\code{NULL} if successful, the error otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
