% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-social.R
\name{FireblazeSocial}
\alias{FireblazeSocial}
\title{Social}
\description{
Use social sites for authentication.
}
\section{Super class}{
\code{\link[fireblaze:Fireblaze]{fireblaze::Fireblaze}} -> \code{FireblazeSocial}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_scope}{\code{FireblazeSocial$set_scope()}}
\item \href{#method-launch_google}{\code{FireblazeSocial$launch_google()}}
\item \href{#method-launch_github}{\code{FireblazeSocial$launch_github()}}
\item \href{#method-launch_facebook}{\code{FireblazeSocial$launch_facebook()}}
\item \href{#method-launch_twitter}{\code{FireblazeSocial$launch_twitter()}}
\item \href{#method-clone}{\code{FireblazeSocial$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_in}{\code{fireblaze::Fireblaze$get_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_up">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_up}{\code{fireblaze::Fireblaze$get_signed_up()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="initialize">}\href{../../fireblaze/html/Fireblaze.html#method-initialize}{\code{fireblaze::Fireblaze$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="set_language_code">}\href{../../fireblaze/html/Fireblaze.html#method-set_language_code}{\code{fireblaze::Fireblaze$set_language_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="sign_out">}\href{../../fireblaze/html/Fireblaze.html#method-sign_out}{\code{fireblaze::Fireblaze$sign_out()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_scope"></a>}}
\if{latex}{\out{\hypertarget{method-set_scope}{}}}
\subsection{Method \code{set_scope()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$set_scope(scope)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{Google scope.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Define the scope to request from Google.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch_google"></a>}}
\if{latex}{\out{\hypertarget{method-launch_google}{}}}
\subsection{Method \code{launch_google()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$launch_google(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Google.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch_github"></a>}}
\if{latex}{\out{\hypertarget{method-launch_github}{}}}
\subsection{Method \code{launch_github()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$launch_github(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Github.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch_facebook"></a>}}
\if{latex}{\out{\hypertarget{method-launch_facebook}{}}}
\subsection{Method \code{launch_facebook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$launch_facebook(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Facebook.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch_twitter"></a>}}
\if{latex}{\out{\hypertarget{method-launch_twitter}{}}}
\subsection{Method \code{launch_twitter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$launch_twitter(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Facebook.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeSocial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
