% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ui.R
\name{FireblazeUI}
\alias{FireblazeUI}
\title{Fireblaze}
\description{
Use fireblaze to manage authentications.
}
\section{Super class}{
\code{\link[fireblaze:Fireblaze]{fireblaze::Fireblaze}} -> \code{FireblazeUI}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{signed_in_user}}{Read the signed in user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_providers}{\code{FireblazeUI$set_providers()}}
\item \href{#method-run}{\code{FireblazeUI$run()}}
\item \href{#method-sign_in}{\code{FireblazeUI$sign_in()}}
\item \href{#method-clone}{\code{FireblazeUI$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="initialize">}\href{../../fireblaze/html/Fireblaze.html#method-initialize}{\code{fireblaze::Fireblaze$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="sign_out">}\href{../../fireblaze/html/Fireblaze.html#method-sign_out}{\code{fireblaze::Fireblaze$sign_out()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-signed_in}{\code{fireblaze::Fireblaze$signed_in()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_providers"></a>}}
\if{latex}{\out{\hypertarget{method-set_providers}{}}}
\subsection{Method \code{set_providers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeUI$set_providers(
  google = FALSE,
  facebook = FALSE,
  twitter = FALSE,
  github = FALSE,
  email = FALSE,
  phone = FALSE,
  anonymous = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{google, facebook, twitter, github, email, phone, anonymous}}{Set to \code{TRUE} the providers you want to use, at least one.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Define signin and login providers.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeUI$run(flow = c("popup", "redirect"), helper = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{The signin flow to use, popup or redirect.}

\item{\code{helper}}{Wether to use accountchooser.com upon signing in or signing up with email,
the user will be redirected to the accountchooser.com website and will be able to select
one of their saved accounts. You can disable it by specifying the value below.}

\item{\code{...}}{Any other option to pass to Firebase UI.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Setup the signin form.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign_in"></a>}}
\if{latex}{\out{\hypertarget{method-sign_in}{}}}
\subsection{Method \code{sign_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeUI$sign_in()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get Signed in User Info
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeUI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
