% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-store.R
\name{Firestore}
\alias{Firestore}
\title{FireStore}
\description{
Interact with firestore
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Firestore$new()}}
\item \href{#method-collection}{\code{Firestore$collection()}}
\item \href{#method-get}{\code{Firestore$get()}}
\item \href{#method-add}{\code{Firestore$add()}}
\item \href{#method-capture}{\code{Firestore$capture()}}
\item \href{#method-clone}{\code{Firestore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collection"></a>}}
\if{latex}{\out{\hypertarget{method-collection}{}}}
\subsection{Method \code{collection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$collection(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$add(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-capture"></a>}}
\if{latex}{\out{\hypertarget{method-capture}{}}}
\subsection{Method \code{capture()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$capture()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Firestore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
