% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_from_env}
\alias{get_config_from_env}
\title{Config from Environment Variables}
\usage{
get_config_from_env()
}
\description{
Builds the configuration object from environment variables.
}
\keyword{internal}
