% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_from_file}
\alias{get_config_from_file}
\title{Config from Environment File}
\usage{
get_config_from_file(path = "firebase.rds")
}
\description{
Builds the configuration object from config file.
}
\keyword{internal}
