% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch_social}
\alias{launch_social}
\title{Launch Social Providers}
\usage{
launch_social(super, social, flow = c("popup", "redirect"))
}
\arguments{
\item{super}{Super class \¢ode{\link{Fireblaze}}.}

\item{social}{Social site, google, facebook, twitter, or github.}

\item{flow}{Flow to use, popup or redirect.}
}
\description{
Convenience function to "send" data to JavaScript; launch social providers.
}
\keyword{internal}
