% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_file}
\alias{pkg_file}
\title{Get Path}
\usage{
pkg_file(path)
}
\arguments{
\item{path}{Path to file in \code{inst} directory.}
}
\value{
Pkg path.
}
\description{
Get path to internal file.
}
\keyword{internal}
