% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\alias{config_exists}
\title{Config Internal}
\usage{
read_config(path)

config_exists(path)
}
\arguments{
\item{path}{Path to config file}
}
\description{
Read and checks config.
}
\keyword{internal}
