% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-phone.R
\name{recaptcha}
\alias{recaptcha}
\alias{recaptchaUI}
\title{Recaptcha}
\usage{
recaptchaUI(
  ns = function(x) {
     x
 }
)
}
\arguments{
\item{ns}{Namespace, optional, only required if using
this function in multiple places.}
}
\description{
Add the recaptcha, require for the \code{FirebasePhone}
class.
}
