% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{reqSignout}
\alias{reqSignout}
\title{Requires Signout}
\usage{
reqSignout(...)
}
\arguments{
\item{...}{Any valid \link[shiny]{tags}.}
}
\description{
Define UI element that requires \emph{no} user to be signed in. This
will hide them \emph{viusally} if no user is signed in.
Note that this is not secure as someone can easily change the
CSS when visiting the page to reveal those elements.
}
\seealso{
\code{\link{reqSignin}}
}
