% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-analytics.R
\name{Analytics}
\alias{Analytics}
\title{Analytics}
\description{
Analytics

Analytics
}
\section{Super class}{
\code{\link[firebase:Firebase]{firebase::Firebase}} -> \code{Analytics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analytics$new()}}
\item \href{#method-log_event}{\code{Analytics$log_event()}}
\item \href{#method-set_user_properties}{\code{Analytics$set_user_properties()}}
\item \href{#method-clone}{\code{Analytics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="print">}\href{../../firebase/html/Firebase.html#method-print}{\code{firebase::Firebase$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analytics$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Initialise

Initialise an analytics object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_event"></a>}}
\if{latex}{\out{\hypertarget{method-log_event}{}}}
\subsection{Method \code{log_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analytics$log_event(event, params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{Event to log choose from
\href{https://support.google.com/analytics/answer/9234069}{this list}
of supported events.}

\item{\code{params}}{Event parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Log Event

Log an event.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_user_properties"></a>}}
\if{latex}{\out{\hypertarget{method-set_user_properties}{}}}
\subsection{Method \code{set_user_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analytics$set_user_properties(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments defining the properties of the user.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set user properties
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analytics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
