% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{firebase_config}
\title{Config}
\usage{
firebase_config(
  api_key,
  project_id,
  auth_domain = NULL,
  app_id = NULL,
  measurement_id = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{api_key}{API key of your project.}

\item{project_id}{Id of your web project.}

\item{auth_domain}{Authentication domain, if omitted uses,
attempts to build firebase's default domain.}

\item{app_id}{Application ID.}

\item{measurement_id}{Measurement ID.}

\item{overwrite}{Whether to overwrite any existing configuration file.}
}
\value{
Path to file.
}
\description{
Create the configuration file necessary to running fireblaze.
Note that if you changed the project you must use said ID
here, not the one originally created by Google.
}
\details{
Do not share this file with anyone.
}
\examples{
\dontrun{firebase_config("xXxxx", "my-project")}

}
