% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{useFirebase}
\alias{useFirebaseUI}
\alias{firebaseUIContainer}
\title{Dependencies}
\usage{
useFirebase(analytics = FALSE, firestore = FALSE)

useFirebaseUI(...)

firebaseUIContainer()
}
\arguments{
\item{analytics}{Deprecated. Whether to include analytics.}

\item{firestore}{Whether to include firestore.}

\item{...}{Ignored, for backwards compatibility.}
}
\description{
Include dependencies in your Shiny application.
\code{use_firebase} \emph{must} be included in
every application.
}
\details{
Place \code{useFirebaseUI} \emph{where} you want
the pre-built UI to be placed. Otherwise one
}
\section{Functions}{

\itemize{
\item{\code{useFirebase} Is required for every app that uses this package}
\item{\code{useFirebaseUI} Is required for applications that use \code{\link{FirebaseUI}}}
\item{\code{firebaseUIContainer} To place the container of the pre-built UI where desired}
}
}

