% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{useFirebase}
\alias{useFirebaseUI}
\title{Dependencies}
\usage{
useFirebase(analytics = TRUE, firestore = FALSE)

useFirebaseUI()
}
\arguments{
\item{analytics}{Whether to include analytics.}

\item{firestore}{Whether to include firestore.}
}
\description{
Include dependencies in your Shiny application.
\code{use_firebase} \emph{must} be included in
every application.
}
\section{Functions}{

\itemize{
\item{\code{useFirebase} Is required for every app that uses this package}
\item{\code{useFirebaseUI} Is required for applications that use \code{\link{FirebaseUI}}}
}
}

