% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-social.R
\name{FirebaseSocial}
\alias{FirebaseSocial}
\title{Social}
\description{
Use social sites for authentication.
}
\examples{
library(shiny)
library(firebase)

# define signin
signin <- modalDialog(
  title = "Login",
  actionButton("google", "Google", icon = icon("google"), class = "btn-danger"),
  actionButton("github", "Github", icon = icon("github")),
  footer = NULL
)

ui <- fluidPage(
  useFirebase()
)

server <- function(input, output) {
  showModal(signin)

  f <- FirebaseSocial$new()

  observeEvent(input$google, {
    f$launch_google()
  })

  observeEvent(input$github, {
    f$launch_github()
  })
}

\dontrun{shinyApp(ui, server)}

}
\section{Super class}{
\code{\link[firebase:Firebase]{firebase::Firebase}} -> \code{FirebaseSocial}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FirebaseSocial-set_scope}{\code{FirebaseSocial$set_scope()}}
\item \href{#method-FirebaseSocial-launch_google}{\code{FirebaseSocial$launch_google()}}
\item \href{#method-FirebaseSocial-launch_github}{\code{FirebaseSocial$launch_github()}}
\item \href{#method-FirebaseSocial-launch_facebook}{\code{FirebaseSocial$launch_facebook()}}
\item \href{#method-FirebaseSocial-launch_twitter}{\code{FirebaseSocial$launch_twitter()}}
\item \href{#method-FirebaseSocial-clone}{\code{FirebaseSocial$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="clear"><a href='../../firebase/html/Firebase.html#method-Firebase-clear'><code>firebase::Firebase$clear()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="delete_user"><a href='../../firebase/html/Firebase.html#method-Firebase-delete_user'><code>firebase::Firebase$delete_user()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_access_token"><a href='../../firebase/html/Firebase.html#method-Firebase-get_access_token'><code>firebase::Firebase$get_access_token()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_delete_user"><a href='../../firebase/html/Firebase.html#method-Firebase-get_delete_user'><code>firebase::Firebase$get_delete_user()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_id_token"><a href='../../firebase/html/Firebase.html#method-Firebase-get_id_token'><code>firebase::Firebase$get_id_token()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_sign_out"><a href='../../firebase/html/Firebase.html#method-Firebase-get_sign_out'><code>firebase::Firebase$get_sign_out()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_signed_in"><a href='../../firebase/html/Firebase.html#method-Firebase-get_signed_in'><code>firebase::Firebase$get_signed_in()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_signed_up"><a href='../../firebase/html/Firebase.html#method-Firebase-get_signed_up'><code>firebase::Firebase$get_signed_up()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="initialize"><a href='../../firebase/html/Firebase.html#method-Firebase-initialize'><code>firebase::Firebase$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="is_signed_in"><a href='../../firebase/html/Firebase.html#method-Firebase-is_signed_in'><code>firebase::Firebase$is_signed_in()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="print"><a href='../../firebase/html/Firebase.html#method-Firebase-print'><code>firebase::Firebase$print()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="req_sign_in"><a href='../../firebase/html/Firebase.html#method-Firebase-req_sign_in'><code>firebase::Firebase$req_sign_in()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="req_sign_out"><a href='../../firebase/html/Firebase.html#method-Firebase-req_sign_out'><code>firebase::Firebase$req_sign_out()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="request_id_token"><a href='../../firebase/html/Firebase.html#method-Firebase-request_id_token'><code>firebase::Firebase$request_id_token()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="set_language_code"><a href='../../firebase/html/Firebase.html#method-Firebase-set_language_code'><code>firebase::Firebase$set_language_code()</code></a></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="sign_out"><a href='../../firebase/html/Firebase.html#method-Firebase-sign_out'><code>firebase::Firebase$sign_out()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-set_scope"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-set_scope}{}}}
\subsection{Method \code{set_scope()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$set_scope(scope)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scope}}{Google scope.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Define the scope to request from Google.
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-launch_google"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-launch_google}{}}}
\subsection{Method \code{launch_google()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$launch_google(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Google.
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-launch_github"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-launch_github}{}}}
\subsection{Method \code{launch_github()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$launch_github(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Github.
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-launch_facebook"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-launch_facebook}{}}}
\subsection{Method \code{launch_facebook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$launch_facebook(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Facebook.
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-launch_twitter"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-launch_twitter}{}}}
\subsection{Method \code{launch_twitter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$launch_twitter(flow = c("popup", "redirect"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{Authentication flow, either popup or redirect.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Launch sign in with Facebook.
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirebaseSocial-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FirebaseSocial-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseSocial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
