% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ui.R
\name{FirebaseUI}
\alias{FirebaseUI}
\title{Prebuilt UI}
\description{
Use fireblaze to manage authentications.
}
\section{Super class}{
\code{\link[firebase:Firebase]{firebase::Firebase}} -> \code{FirebaseUI}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tos_url}}{URL to the Terms of Service page.}

\item{\code{privacy_policy_url}}{The URL to the Privacy Policy page.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_providers}{\code{FirebaseUI$set_providers()}}
\item \href{#method-set_tos_url}{\code{FirebaseUI$set_tos_url()}}
\item \href{#method-set_privacy_policy_url}{\code{FirebaseUI$set_privacy_policy_url()}}
\item \href{#method-launch}{\code{FirebaseUI$launch()}}
\item \href{#method-clone}{\code{FirebaseUI$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_signed_in">}\href{../../firebase/html/Firebase.html#method-get_signed_in}{\code{firebase::Firebase$get_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="get_signed_up">}\href{../../firebase/html/Firebase.html#method-get_signed_up}{\code{firebase::Firebase$get_signed_up()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="initialize">}\href{../../firebase/html/Firebase.html#method-initialize}{\code{firebase::Firebase$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="is_signed_in">}\href{../../firebase/html/Firebase.html#method-is_signed_in}{\code{firebase::Firebase$is_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="req_signed_in">}\href{../../firebase/html/Firebase.html#method-req_signed_in}{\code{firebase::Firebase$req_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="set_language_code">}\href{../../firebase/html/Firebase.html#method-set_language_code}{\code{firebase::Firebase$set_language_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="sign_out">}\href{../../firebase/html/Firebase.html#method-sign_out}{\code{firebase::Firebase$sign_out()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_providers"></a>}}
\if{latex}{\out{\hypertarget{method-set_providers}{}}}
\subsection{Method \code{set_providers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseUI$set_providers(
  google = FALSE,
  facebook = FALSE,
  twitter = FALSE,
  github = FALSE,
  email = FALSE,
  phone = FALSE,
  anonymous = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{google, facebook, twitter, github, email, phone, anonymous}}{Set to \code{TRUE} the providers you want to use, at least one.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Define signin and login providers.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_tos_url"></a>}}
\if{latex}{\out{\hypertarget{method-set_tos_url}{}}}
\subsection{Method \code{set_tos_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseUI$set_tos_url(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Defines Tterms of Services URL
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_privacy_policy_url"></a>}}
\if{latex}{\out{\hypertarget{method-set_privacy_policy_url}{}}}
\subsection{Method \code{set_privacy_policy_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseUI$set_privacy_policy_url(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Defines Privacy Policy URL
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\if{latex}{\out{\hypertarget{method-launch}{}}}
\subsection{Method \code{launch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseUI$launch(flow = c("popup", "redirect"), account_helper = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flow}}{The signin flow to use, popup or redirect.}

\item{\code{account_helper}}{Wether to use accountchooser.com upon signing in or signing up with email,
the user will be redirected to the accountchooser.com website and will be able to select
one of their saved accounts. You can disable it by specifying the value below.}

\item{\code{...}}{Any other option to pass to Firebase UI.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Setup the signin form.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirebaseUI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
