% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-core.R
\name{Fireblaze}
\alias{Fireblaze}
\title{Fireblaze}
\description{
Use fireblaze to manage authentications.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session}}{A valid Shiny session.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{signed_in}}{Read the signed in user.}

\item{\code{signed_up}}{Read the signed in user.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fireblaze$new()}}
\item \href{#method-sign_out}{\code{Fireblaze$sign_out()}}
\item \href{#method-get_signed_in}{\code{Fireblaze$get_signed_in()}}
\item \href{#method-get_signed_up}{\code{Fireblaze$get_signed_up()}}
\item \href{#method-set_language_code}{\code{Fireblaze$set_language_code()}}
\item \href{#method-clone}{\code{Fireblaze$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$new(
  persistence = c("session", "none", "local"),
  config_path = "fireblaze.rds",
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{persistence}}{How the auth should persit: \code{none}, the user has to sign in at every visit,
\code{session} will only persist in current tab, \code{local} persist even when window is closed.}

\item{\code{config_path}}{Path to the configuration file as created by \code{\link{create_config}}.}

\item{\code{session}}{A valid shiny session.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise Fireblaze
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign_out"></a>}}
\if{latex}{\out{\hypertarget{method-sign_out}{}}}
\subsection{Method \code{sign_out()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$sign_out()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Signs user out
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_signed_in"></a>}}
\if{latex}{\out{\hypertarget{method-get_signed_in}{}}}
\subsection{Method \code{get_signed_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$get_signed_in()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Signed in user details triggered when auth states changes
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_signed_up"></a>}}
\if{latex}{\out{\hypertarget{method-get_signed_up}{}}}
\subsection{Method \code{get_signed_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$get_signed_up()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Signed up user details triggered explicitely by user (e.g.: on click)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_language_code"></a>}}
\if{latex}{\out{\hypertarget{method-set_language_code}{}}}
\subsection{Method \code{set_language_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$set_language_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{iso639-1 language code.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set language code for auth provider
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fireblaze$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
