% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-email-password.R
\name{FireblazeEmailPassword}
\alias{FireblazeEmailPassword}
\title{Email & Password}
\description{
Use fireblaze to manage authentications.
}
\section{Super class}{
\code{\link[fireblaze:Fireblaze]{fireblaze::Fireblaze}} -> \code{FireblazeEmailPassword}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{created}}{Results of account creation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-create}{\code{FireblazeEmailPassword$create()}}
\item \href{#method-sign_in}{\code{FireblazeEmailPassword$sign_in()}}
\item \href{#method-get_created}{\code{FireblazeEmailPassword$get_created()}}
\item \href{#method-clone}{\code{FireblazeEmailPassword$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_in}{\code{fireblaze::Fireblaze$get_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="get_signed_up">}\href{../../fireblaze/html/Fireblaze.html#method-get_signed_up}{\code{fireblaze::Fireblaze$get_signed_up()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="initialize">}\href{../../fireblaze/html/Fireblaze.html#method-initialize}{\code{fireblaze::Fireblaze$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="is_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-is_signed_in}{\code{fireblaze::Fireblaze$is_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="req_signed_in">}\href{../../fireblaze/html/Fireblaze.html#method-req_signed_in}{\code{fireblaze::Fireblaze$req_signed_in()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="set_language_code">}\href{../../fireblaze/html/Fireblaze.html#method-set_language_code}{\code{fireblaze::Fireblaze$set_language_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fireblaze" data-topic="Fireblaze" data-id="sign_out">}\href{../../fireblaze/html/Fireblaze.html#method-sign_out}{\code{fireblaze::Fireblaze$sign_out()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$create(email, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email, password}}{Credentials as entered by the user.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sign in with create
}

\subsection{Returns}{
\code{NULL} if successful, the error otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign_in"></a>}}
\if{latex}{\out{\hypertarget{method-sign_in}{}}}
\subsection{Method \code{sign_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$sign_in(email, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email, password}}{Credentials as entered by the user.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sign in with email
}

\subsection{Returns}{
\code{NULL} if successful, the error otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_created"></a>}}
\if{latex}{\out{\hypertarget{method-get_created}{}}}
\subsection{Method \code{get_created()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$get_created()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get account creation results
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FireblazeEmailPassword$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
