% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{reqSignin}
\alias{reqSignin}
\title{Requires Signin}
\usage{
reqSignin(...)
}
\arguments{
\item{...}{Any valid \link[shiny]{tags}.}
}
\description{
Define UI element that require the user to be signed in. This
will hide them \emph{viusally} until the user signs in.
Note that this is not secure as someone can easily change the
CSS when visiting the page to reveal those elements.
}
