% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-store.R
\name{Store}
\alias{Store}
\title{Store}
\value{
An object of class \code{Store}.
}
\description{
Store

Store
}
\examples{

## ------------------------------------------------
## Method `Store$query`
## ------------------------------------------------

\dontrun{
s <- Storage$new()

s$
  query("users", "where('age', '>', 18)", id = "adults", observe = TRUE)

s$add(.collection = "users", name = "John", age = 28)
s$add(.collection = "users", name = "Bob", age = 12)

observeEvent(input$adults, {
  # triggered wehn a user with age > 18 is added, changed or deleted
})

s$unsubscribe("adults")
}
}
\section{Super class}{
\code{\link[firebase:Firebase]{firebase::Firebase}} -> \code{Store}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Store-new}{\code{Store$new()}}
\item \href{#method-Store-query}{\code{Store$query()}}
\item \href{#method-Store-add}{\code{Store$add()}}
\item \href{#method-Store-set}{\code{Store$set()}}
\item \href{#method-Store-unsubscribe}{\code{Store$unsubscribe()}}
\item \href{#method-Store-clone}{\code{Store$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="expose_app"><a href='../../firebase/html/Firebase.html#method-Firebase-expose_app'><code>firebase::Firebase$expose_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="firebase" data-topic="Firebase" data-id="print"><a href='../../firebase/html/Firebase.html#method-Firebase-print'><code>firebase::Firebase$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-new"></a>}}
\if{latex}{\out{\hypertarget{method-Store-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$new(
  config_path = "firebase.rds",
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_path}}{Path to the configuration file as created by \code{\link{firebase_config}}.}

\item{\code{session}}{A valid shiny session.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialiases Firebase Storage

Initialises the Firebase Storage application client-side.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-query"></a>}}
\if{latex}{\out{\hypertarget{method-Store-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$query(collection, query, id, observe = FALSE, cb = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{Name of the collection to execute the query on.}

\item{\code{id}}{The \code{id} is also used to unsubscribe from the query and get the results
of the query server-side, see \code{observe} argument.}

\item{\code{observe}}{If set to \code{TRUE} the client will listen to the results of the query and
and send the results to the server at the given input \code{id}.}

\item{\code{cb}}{A callback function to execute on the client side when the results
are available. This is also JavaScript code as a string that is evaluated on the client side,
namely a function that accepts the results as an argument, this is passed to \code{onSnapshot}.}

\item{\code{response}}{query The query to execute, this is expecting JavaScript code
as a string that is evaluated on the client side, e.g.: \code{where("age", ">", 18)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Query

Query a collection in the Firestore database.


You should normally not have to listen the result of individual queries,
as this will likely be inefficient. Instead, you should listen to the collection
or a small subset of the collection in on place.
This is because firebase is by default a real-time database, any changes to the
subset of documents defined by the query will trigger the callback.
For instance, listening to the results of \code{where("age", ">", 18)} and \code{where("age", ">", 16)}
will trigger the callback twice when a user with age 17 is added.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
s <- Storage$new()

s$
  query("users", "where('age', '>', 18)", id = "adults", observe = TRUE)

s$add(.collection = "users", name = "John", age = 28)
s$add(.collection = "users", name = "Bob", age = 12)

observeEvent(input$adults, {
  # triggered wehn a user with age > 18 is added, changed or deleted
})

s$unsubscribe("adults")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-add"></a>}}
\if{latex}{\out{\hypertarget{method-Store-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$add(..., .collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named list of values to add as a document.}

\item{\code{.collection}}{Name of the collection to add the document to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add a document to a collection
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-set"></a>}}
\if{latex}{\out{\hypertarget{method-Store-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$set(..., .doc, .collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named list of values to set as a document.}

\item{\code{.doc}}{Name of the document to create or update.}

\item{\code{.collection}}{Name of the collection to add the document to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create or update a document
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-Store-unsubscribe}{}}}
\subsection{Method \code{unsubscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$unsubscribe(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The \code{id} of the query to unsubscribe from.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Unsubscribe from a query
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Store-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
