% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{use_fireblaze}
\alias{use_fireblaze}
\title{Dependencies}
\usage{
use_fireblaze(ui = FALSE, analytics = TRUE, firestore = FALSE)
}
\arguments{
\item{ui}{Wehther to include the UI, required for \code{\link{FireblazeUI}}.}

\item{analytics}{Whether to include analytics.}

\item{firestore}{Whether to include firestore.}
}
\description{
Include dependencies in your Shiny application.
This \emph{must} be included at the \emph{bottom} the application.
}
