% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{use_fireblaze}
\alias{use_fireblaze}
\title{Dependencies}
\usage{
use_fireblaze()
}
\description{
Include dependencies in your Shiny application.
This \emph{must} be included at the \emph{bottom} the application.
}
